prod/lambda/return_table/requirements.txt
mysql-connector-python==8.0.33